+++ src/singletons/WindowManager.hpp
@@ -107,6 +107,9 @@ private:
     // Apply a window layout for this window manager.
     void applyWindowLayout(const WindowLayout &layout);
 
+    // For the midnight timer
+    int getNextMidnight();
+
     // Contains the full path to the window layout file, e.g. /home/pajlada/.local/share/Chatterino/Settings/window-layout.json
     const QString windowLayoutFilePath;
 
@@ -126,6 +129,7 @@ private:
     pajlada::SettingListener wordFlagsListener_;
 
     QTimer *saveTimer;
+    QTimer *midnightTimer;
     QTimer miscUpdateTimer_;
 };